app/src/main/java/com/sample/testingdemo/LoginViewModel.kt
package com.sample.testingdemo

import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import dagger.hilt.android.lifecycle.HiltViewModel

import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class LoginViewModel @Inject constructor(private val repository: LoginRepository) : ViewModel() {

    private val _loginState = MutableStateFlow(LoginState())
    val loginState: StateFlow<LoginState> = _loginState

    fun login(username: String, password: String) {
        viewModelScope.launch {
            repository.login(username, password).collect { result ->
                _loginState.value = LoginState(isSuccess = result.getOrNull()!!)
            }
        }
    }
}

data class LoginState(
    val isSuccess: Boolean = false
)